chore: Update job board widget to use new components
