+++ app/assets/jobs.json
@@ -1,0 +1,26 @@
+[
+  {
+    "company": "株式会社ABCDEFGHIJKLMNOPQRSTUVWXYZABCDEFGHIJKLMNOPQRSTUVWXYZ",
+    "title": "Remix デベロッパー",
+    "tags": ["フルリモート", "業務委託OK", "週3日以下OK", "B2B", "SPA"],
+    "openAt": "2024-07-01 15:00",
+    "expiredAt": "2024-09-01 15:00",
+    "href": "https://www.techtalk.jp/"
+  },
+  {
+    "company": "株式会社ふがふが",
+    "title": "プロダクトエンジニア",
+    "tags": ["B2C", "リモート", "Full Stack"],
+    "openAt": "2024-07-01 15:00",
+    "expiredAt": "2024-08-01 15:00",
+    "href": "https://www.techtalk.jp/"
+  },
+  {
+    "company": "株式会社TechTalk",
+    "title": "Remix デベロッパー",
+    "tags": ["フルリモート", "業務委託", "週3日以下OK", "B2B", "Full Stack"],
+    "openAt": "2024-07-01 15:00",
+    "expiredAt": "2024-08-01 14:59",
+    "href": "https://www.techtalk.jp/"
+  }
+]
